Create new folder for files related to article form and record initialization
